cleanup lexer and parser
